[#27] Add color utils and storage utils
