+++ pages/character.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useRef } from 'react';
+import React, {useEffect, useRef, useState} from 'react';
 import Layout from '../app/layout';
 import SelectLayout from "@/app/selectLayout";
 import ActionButton from "@/app/components/actionButton/ActionButton";
@@ -6,69 +6,68 @@ import SelectItems from "@/app/components/selectItems/selectItems";
 import "./character.css";
 import "../app/globals.css";
 import Image from "next/image";
-import { getStoredValue, setStoredValue, removeStoredValue } from '@/app/utils/storageUtils';
-import { getColorWord, getColors } from '@/app/utils/colorUtils';
+import {getStoredValue, setStoredValue} from '@/app/utils/storageUtils';
+import {getHairColors} from '@/app/utils/colorUtils';
 
 const Character: React.FC = () => {
+    const isClient = typeof window !== 'undefined';
     const [name, setName] = useState<string>('');
     const [selectedImageType, setSelectedImageType] = useState<string>('girl');
     const [selectedColorWord, setSelectedColorWord] = useState<string>('dark');
     const nameInputRef = useRef<HTMLInputElement>(null);
 
     useEffect(() => {
-        const storedName = getStoredValue('characterName');
-        if (storedName) {
-            setName(storedName);
-        }
+        if (isClient) {
+            const storedName = getStoredValue('characterName');
+            if (storedName) {
+                setName(storedName);
+            }
 
-        const storedImageType = getStoredValue('selectedImageType');
-        if (storedImageType) {
-            setSelectedImageType(storedImageType);
-        }
+            const storedImageType = getStoredValue('selectedImageType');
+            if (storedImageType) {
+                setSelectedImageType(storedImageType);
+            }
 
-        const storedColorWord = getStoredValue('selectedColorWord');
-        if (storedColorWord) {
-            setSelectedColorWord(storedColorWord);
+            const storedColorWord = getStoredValue('selectedHairColorWord');
+            if (storedColorWord) {
+                setSelectedColorWord(storedColorWord);
+            }
         }
-    }, []);
+    }, [isClient]);
 
     const handleNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         const newName = e.target.value;
         setName(newName);
-        setStoredValue('characterName', newName);
+        if (isClient) {
+            setStoredValue('characterName', newName);
+        }
     };
 
     const handleImageClick = (type: string) => {
         setSelectedImageType(type);
-        setStoredValue('selectedImageType', type);
+        if (isClient) {
+            setStoredValue('selectedImageType', type);
+        }
     };
 
     const handleColorClick = (index: number) => {
-        const colorWord = getColors()[index].word;
+        const colorWord = getHairColors()[index].word;
         setSelectedColorWord(colorWord);
-        setStoredValue('selectedColorWord', colorWord);
-    };
-
-    const clearLocalStorage = () => {
-        removeStoredValue('selectedImageType');
-        removeStoredValue('selectedColorWord');
-        setSelectedImageType('girl');
-        setSelectedColorWord('dark');
+        if (isClient) {
+            setStoredValue('selectedHairColorWord', colorWord);
+        }
     };
 
     const renderLeftChildren = () => {
-        const imageType = getStoredValue('selectedImageType') || selectedImageType;
-        const colorWord = getStoredValue('selectedColorWord') || selectedColorWord;
         return (
             <div className="h-full w-auto flex justify-center items-center">
-                <Image className="character-image" src={`/images/${imageType}_${colorWord}.svg`}
+                <Image className="character-image" src={`/images/${selectedImageType}_${selectedColorWord}.svg`}
                        alt="Ausgewählter Charakter" width={400} height={627} />
             </div>
         );
     };
 
     const renderRightChildren = () => {
-        const colorWord = getStoredValue('selectedColorWord') || selectedColorWord;
         return (
             <Layout>
                 <div>
@@ -76,13 +75,13 @@ const Character: React.FC = () => {
                     <h2 className="font-bold mb-4">Geschlecht</h2>
                     <div className="text-center">
                         <SelectItems onClick={index => handleImageClick(index === 0 ? 'girl' : 'boy')} images={[
-                            { src: `/images/girl_${colorWord}.svg`, desc: "Weiblich" },
-                            { src: `/images/boy_${colorWord}.svg`, desc: "Männlich" },
+                            { src: `/images/girl_${selectedColorWord}.svg`, desc: "Weiblich" },
+                            { src: `/images/boy_${selectedColorWord}.svg`, desc: "Männlich" },
                         ]} />
                     </div>
                     <h2 className="font-bold mb-4">Haarfarbe</h2>
                     <div className="text-center">
-                        <SelectItems onClick={handleColorClick} colorCodes={getColors().map(color => color.code)} />
+                        <SelectItems onClick={handleColorClick} colorCodes={getHairColors().map(color => color.code)} />
                     </div>
                     <div>
                         <div>
@@ -104,7 +103,7 @@ const Character: React.FC = () => {
             </Layout>
         );
     };
-
+    //TODO Routing
     let rocketpage;
     return (
         <div>
